backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
langchain==0.1.0
langchain-text-splitters==0.0.1
PyPDF2==3.0.1
python-docx==1.1.0
supabase==2.3.0
google-generativeai==0.3.1
python-dotenv==1.0.0
